the first simple test to validate we can connect.
We decided to just write separate, simplified tests for the basics,
making async support somewhat of a second class citizen, but it's a
valid first step. With the right code layout, one should be able
to reuse all the blocking code where it matters.

This may also make us push code down into `git-protocol`, where
it probably should be anyway. Let's see if this is feasible.