migrate independent parts of ref parsing to generalized handshake.
V2 ref fetching (invoking a command) is indeed only available in V2
which in turn is only for fetch.

Maybe the solution here is rather to make ls-refs a general command,
which might lead to reorganzation. For now, it's OK to keep it there
I think knowing push is only V1 which is what this is for anyway.