[git-object] Encode empty tags like git does
A tag with an empty commit message is now displayed with an extra
newline, consitently with the output of `git cat-file`.

The test fixture for `empty.txt` was updated, which means that the
`round_trip` test now expects a correct round trip with the new format.

Note that the changes to tests in `immutable` reflect the new state of
the code:
* `TagRefIter` now parses this body with an empty message.
* Where the length of the data was divided by 2, the point was to create
  an inconsistent dataset leading to an error. Unfortunately, with the
  extra character, it lines up so that the message can be parsed, so I
  had to change the magic number to keep the previous property.

Fixes #603