fix: Discover repo with relative path and ceiling
A couple of problems are repaired to allow discovering a repository from
"." with a ceiling directory.

One problem was that find_ceiling_height() did the wrong thing when
confronted with any relative search_dir. This is resolved by converting
search_dir to be absolute if it is relative.

The other problem was that discover_opts() also mishandled relative paths.
When the cursor started out as ".", cursor.pop() would be blindly called
such that cursor would be "" for the second iteration. When a ceiling
directory was in use such that there was a max_height, the current height
would be burned going from ".", to "", and then to "<cwd>", before finally
actually getting to a real parent directory. This problem is ameliorated by
testing whether the cursor has a non-empty parent before popping.

N.B. the new test case relies on the test running from the git-discover
directory such that the gitoxide repository will be found. This is a bit
fragile and will fail if, for example, the test is run from an unpacked
gitoxide source tarball.