feat: `Remote` knows about its `tagOpt` configuration.
That way it's clear if it should or shouldn't fetch included/reachable
tags automatically.

The default setting for this is to include tags, similar to `git`.

The `fetch_tags()` accessor allows to query this information, and the
`with_fetch_tags()` builder method allows to set the value comfortably
right after creating the `Remote` instance.

The `tagOpt` key will also be written as part of the remote's git
configuration.

Clone operations can set the `Tags` setting when configuring the
remote in a callback.

This also comes with a fix to assure that ref-updates aren't skipped
just because there was no pack to receive. That way, locally missing
refs or tags will automatically be put back.