Assure recursive shared indices (possibly malicious) don't crash us
The way this is resolved is similar to what git does by means of comparing
the expected and actual hashes of the file at hand.

Note that we do not actually protect against recursion but assume that
nobody can manufacture a file hash that in fact is recursive. We know that
this can actually happen but chose to not consider it worth protecting against
just yet, as seemingly neither does git.