fix: don't enforce V2 as protocol, but smoothly downgrade like git does.
For backward compatibility the shared handshake implementation allows the
transport to control which protocol versions it wants to support
to allow optiomizing for one special case, namely to prevent it to
read all V1 refs on old servers but abort instead, closing the connection
without delay.

Now we leave this feature for custom transports (who usually come with custom
servers) and instead support fallbacks to other protocols if the server
demands it.