feat!: make streaming otional for any reqwest.
One can now indicate when initiating a reqwest that the transport doesn't
have to stream the data, even though it will always be provided to an
`std::io::Write`.

Note that this is at the discretion of the transport implementation and streaming
might still be done despite it not being requested.

Note that the caller should set this 'streaming' flag if the upper bound of data
is high for keeping it in memory or can't be estimated. This is generally true
when sending packs.