doc: explain how it's possible to deal with the first commit when comparing trees (#676)
The reason the other tree isn't an option is that it's a special case that can more easily be handled
with an `.unwrap_or_else(|| repo.empty_tree())` (or similar) for those who need it.

Making the empty tree explicit also helps to deal with diffs from the empty tree (which can't be `Option<Tree>`)
to the first tree of the first commit.

Fixes #676.