feat: Support git default date format
This is the format output by default by `git log` or when using
`--pretty=%ad`.

The new git_date::time::format::GIT_DEFAULT format description may be used
to output date strings in this format. It is also now used by
git_date::parse() to accept date strings that may be in this format.

Refs: #679

Signed-off-by: Peter Grayson <pete@jpgrayson.net>