fix: file:// command invocation won't spill stderr output.
This usually doesn't add any benefit to the user as we might see
events like the git process' failure to flush to a closed channel
even though this is entirely handled by the Rust side of things.

I can imagine that one day this might become a configurable to help
with debugging to help making better-behaved clients, but maybe it
won't ever matter once the default file:// transport is built-in and
native.