fix: Stricter raw date parsing
The raw date parser (git_date::parse::function::parse_raw()) accepted some
inputs that it should not have. Specifically, it would accept:

- Any character for the timezone offset's sign
- Trailing, non-whitespace characters after the timezone offset

Now either '+' or '-' is required for the timezone offset sign and only
trailing whitespace is allowed.

Additional tests are added to cover both acceptable and unacceptable
inputs.

N.B. the raw date parser is still accepting of whitespace leading, in the
middle of, and trailing the date string. A yet stricter parser would only
allow a single space character between the seconds-since-epoch and the
timezone offset.

Signed-off-by: Peter Grayson <pete@jpgrayson.net>